# Reduce allocations by optimizing branching and black holes