@ChrisPenner 2024-12-09
Hrmm, really not sure how this could cause that transcript to fail 🤔